# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Back End Assignment API Documentation",
    "version": "1.0.0",
    "description": "This is the API documentation for the event Management System."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Event": {
        "type": "object",
        "required": [
          "name",
          "date",
          "capacity"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Tech Conference"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-15T10:00:00Z"
          },
          "capacity": {
            "type": "integer",
            "minimum": 5,
            "example": 100
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "cancelled",
              "completed"
            ],
            "default": "active"
          },
          "category": {
            "type": "string",
            "enum": [
              "conference",
              "workshop",
              "meetup",
              "seminar",
              "general"
            ],
            "default": "general"
          },
          "registrationCount": {
            "type": "integer",
            "minimum": 0,
            "example": 0
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/event": {
      "post": {
        "summary": "Create a new event",
        "tags": [
          "Events"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "date",
                  "capacity"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 3,
                    "example": "Tech Conference"
                  },
                  "date": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2026-06-15T10:00:00Z"
                  },
                  "capacity": {
                    "type": "integer",
                    "minimum": 5,
                    "example": 100
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "cancelled",
                      "completed"
                    ],
                    "default": "active"
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "conference",
                      "workshop",
                      "meetup",
                      "seminar",
                      "general"
                    ],
                    "default": "general"
                  },
                  "registrationCount": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          }
        }
      },
      "get": {
        "summary": "Retrieve all events",
        "tags": [
          "Events"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved events",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          }
        }
      }
    },
    "/event/{id}": {
      "get": {
        "summary": "Retrieve a specific event by ID",
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Event ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Event retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID"
          },
          "404": {
            "description": "Event not found"
          }
        }
      },
      "put": {
        "summary": "Update a specific event",
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Event ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "minProperties": 1,
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "date": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "capacity": {
                    "type": "integer"
                  },
                  "category": {
                    "type": "string"
                  },
                  "registrationCount": {
                    "type": "integer"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "cancelled",
                      "completed"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Event not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a specific event",
        "tags": [
          "Events"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Event ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Event deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid ID"
          },
          "404": {
            "description": "Event not found"
          }
        }
      }
    }
  },
  "tags": []
}